+++ packages/components/tabs/src/tab-group.ts
@@ -1,10 +1,10 @@
 import { localized, msg } from '@lit/localize';
 import { type ScopedElementsMap, ScopedElementsMixin } from '@open-wc/scoped-elements/lit-element.js';
-import { Button } from '@sl-design-system/button';
 import { Icon } from '@sl-design-system/icon';
-import { type EventEmitter, RovingTabindexController, anchor, event, isPopoverOpen } from '@sl-design-system/shared';
+import { MenuButton, MenuItem } from '@sl-design-system/menu';
+import { type EventEmitter, RovingTabindexController, event, getScrollParent } from '@sl-design-system/shared';
 import { type CSSResultGroup, LitElement, type PropertyValues, type TemplateResult, html, nothing } from 'lit';
-import { property, query, queryAssignedElements, state } from 'lit/decorators.js';
+import { property, state } from 'lit/decorators.js';
 import styles from './tab-group.scss.js';
 import { TabPanel } from './tab-panel.js';
 import { Tab } from './tab.js';
@@ -21,9 +21,7 @@ declare global {
 
 export type SlTabChangeEvent = CustomEvent<number>;
 
-export type TabsAlignment = 'start' | 'filled';
-
-let nextUniqueId = 0;
+export type TabsAlignment = 'start' | 'center' | 'end' | 'stretch';
 
 const OBSERVER_OPTIONS: MutationObserverInit = {
   attributes: true,
@@ -32,380 +30,392 @@ const OBSERVER_OPTIONS: MutationObserverInit = {
   attributeOldValue: true
 };
 
+let nextUniqueId = 0;
+
 /**
  * A tab group component that can contain tabs and tab panels.
  *
  * ```html
- *    <sl-tab-group>
- *         <sl-tab selected>First tab</sl-tab>
- *         <sl-tab-panel>Content of the tab 1</sl-tab-panel>
+ *   <sl-tab-group>
+ *     <sl-tab>First tab</sl-tab>
+ *     <sl-tab selected>Second tab</sl-tab>
  *
- *         <sl-tab>Second tab</sl-tab>
- *         <sl-tab-panel>Content of the tab 2</sl-tab-panel>
- *    </sl-tab-group>
+ *     <sl-tab-panel>Content of tab 1</sl-tab-panel>
+ *     <sl-tab-panel>Content of tab 2</sl-tab-panel>
+ *   </sl-tab-group>
  * ```
  *
- * @slot default - a place for the tab group content.
+ * @csspart container - The container for the tabs.
+ * @csspart wrapper - Wraps the scroll container and menu button.
+ * @csspart scroller - The scroll container of the tabs.
+ * @csspart tablist - The tablist element which also contains the active tab indicator
+ * @csspart panels - The container for the tab panels.
+ *
+ * @cssprop --sl-tab-group-menu-min-inline-size - The minimum inline size of the menu.
+ * @cssprop --sl-tab-group-menu-max-inline-size - The maximum inline size of the menu.
+ *
+ * @slot default - Tab panels or other tab content here.
+ * @slot tabs - The tabs to display.
  */
 @localized()
 export class TabGroup extends ScopedElementsMixin(LitElement) {
-  /** @private */
+  /** @internal */
   static get scopedElements(): ScopedElementsMap {
     return {
-      'sl-button': Button,
       'sl-icon': Icon,
+      'sl-menu-button': MenuButton,
+      'sl-menu-item': MenuItem,
       'sl-tab': Tab,
       'sl-tab-panel': TabPanel
     };
   }
 
-  /** @private */
+  /** @internal */
   static override styles: CSSResultGroup = styles;
 
+  /** Unique prefix ID for each component in the light DOM. */
+  #idPrefix = `sl-tab-group-${nextUniqueId++}`;
+
   /**
-   * Unique ID for each tab group component present.
+   * Observe changes to the selected tab and update accordingly. This observer
+   * is necessary for changes to the selected tab that are made programmatically.
+   * Selected changes made by the user are handled by the click event listener.
    */
-  #tabGroupId = `sl-tab-group-${nextUniqueId++}`;
+  #mutationObserver = new MutationObserver(entries => {
+    entries.forEach(entry => {
+      if (entry.attributeName === 'selected' && entry.oldValue === null) {
+        this.#mutationObserver?.disconnect();
 
-  #observer?: MutationObserver;
+        // Update the selected tab with the observer turned off to avoid loops
+        this.#updateSelectedTab(entry.target as Tab);
 
+        this.#mutationObserver?.observe(this, OBSERVER_OPTIONS);
+      }
+    });
+  });
+
+  /**
+   * Observe changes to the size of the tablist so:
+   * - we can determine when to display an overflow menu with tab items
+   * - we know when we need to reposition the active tab indicator
+   */
+  #resizeObserver = new ResizeObserver(() => {
+    this.#shouldAnimate = false;
+    this.#updateSize();
+    this.#shouldAnimate = true;
+  });
+
+  /** Manage keyboard navigation between tabs. */
   #rovingTabindexController = new RovingTabindexController<Tab>(this, {
     focusInIndex: (elements: Tab[]) => elements.findIndex(el => el.selected),
-    elements: () => (isPopoverOpen(this.listbox) ? this.#allTabs : this.tabs) || [],
+    elements: () => this.tabs || [],
     isFocusableElement: (el: Tab) => !el.disabled
   });
 
-  /** All slotted tabs. */
-  #allTabs: Tab[] = [];
+  /** Determines whether the active tab indicator should animate. */
+  #shouldAnimate = false;
 
-  /** Whether more button needs to be shown */
-  #showMore = false;
+  /** The alignment of tabs within the wrapper. */
+  @property({ attribute: 'align-tabs', reflect: true }) alignTabs?: TabsAlignment;
 
-  /** Button used to show all tabs */
-  #moreButton!: HTMLButtonElement;
+  /** The menu items to render when the tabs are overflowing. */
+  @state() menuItems?: Array<{ tab: Tab; disabled?: boolean; title: string; subtitle?: string }>;
 
-  /** Observe the tablist width. */
-  #sizeObserver = new ResizeObserver(() => {
-    this.#updateSelectionIndicator();
-  });
+  /** The currently selected tab. */
+  @state() selectedTab?: Tab;
 
-  /** The slotted tabs. */
-  @queryAssignedElements({ slot: 'tabs' }) tabs?: Tab[];
-
-  /** The current tab node selected in the tab group. */
-  @state() private selectedTab: Tab | null = this.#initialSelectedTab;
-
-  /** The current tab node selected in the tab listbox (dropdown). */
-  @state() private selectedTabInListbox: Tab | null = this.#initialSelectedTab;
+  /** Whether the menu button needs to be shown. */
+  @state() showMenu = false;
 
   /** Emits when the tab has been selected/changed. */
-  @event() tabChange!: EventEmitter<SlTabChangeEvent>;
+  @event({ name: 'sl-tab-change' }) tabChangeEvent!: EventEmitter<SlTabChangeEvent>;
+
+  /** The slotted tabs. */
+  @state() tabPanels?: TabPanel[];
+
+  /** The slotted tabs. */
+  @state() tabs?: Tab[];
 
   /** Renders the tabs vertically instead of the default horizontal  */
   @property({ type: Boolean, reflect: true }) vertical?: boolean;
 
-  /** The alignment of tabs inside sl-tab-group  */
-  @property({ reflect: true }) alignment: TabsAlignment = 'start';
-
-  /** The listbox element with all tabs list. */
-  @query('[popover]') listbox!: HTMLElement;
-
-  /** Get the selected tab button, or the first tab button. */
-  get #initialSelectedTab(): Tab | null {
-    return this.querySelector('sl-tab[selected]') || this.querySelector('sl-tab:not([disabled])');
-  }
-
-  override render(): TemplateResult {
-    this.#moreButton = this.renderRoot.querySelector('#more-btn') as HTMLButtonElement;
-    this.#allTabs = Array.from(this.querySelectorAll('sl-tab'))?.map(tab => tab.cloneNode(true)) as Tab[];
-    this.#allTabs.forEach(tab => tab.classList.add('listbox-tab'));
-
-    return html`
-      <div class="container" part="container" @keydown=${this.#handleKeydown}>
-        <div class="wrapper" part="wrapper">
-          <div @click=${this.#handleTabChange} role="tablist" part="tab-list">
-            <span class="indicator" role="presentation"></span>
-            <slot name="tabs" @slotchange=${() => this.#rovingTabindexController.clearElementCache()}></slot>
-            <div
-              ${anchor({ element: this.#moreButton, position: this.vertical ? 'bottom-start' : 'bottom-end' })}
-              @click=${this.#handleTabChange}
-              @toggle=${this.#onToggle}
-              aria-labelledby="more-btn"
-              id="tabs-popover"
-              part="listbox"
-              popover
-              role="listbox"
-            >
-              ${this.#allTabs}
-            </div>
-          </div>
-        </div>
-        ${this.#showMore
-          ? html`
-              <sl-button
-                @click=${this.#onClick}
-                aria-controls="tabs-popover"
-                aria-expanded="false"
-                aria-haspopup="true"
-                aria-label=${msg('Show all')}
-                id="more-btn"
-                fill="ghost"
-                size="md"
-                variant="primary"
-              >
-                <sl-icon name="ellipsis"></sl-icon>
-              </sl-button>
-            `
-          : nothing}
-      </div>
-      <slot></slot>
-    `;
-  }
-
   override connectedCallback(): void {
     super.connectedCallback();
 
-    this.#sizeObserver.observe(this);
-    this.#updateSlots();
+    this.#mutationObserver.observe(this, OBSERVER_OPTIONS);
+
+    // We need to wait for the next frame so the element has time to render
+    requestAnimationFrame(() => {
+      const tablist = this.renderRoot.querySelector('[part="tablist"]') as Element;
+
+      // We want to observe the size of the tablist, not the
+      // container or wrapper. The tablist is the element that
+      // changes size for example when fonts are loaded. The
+      // other elements do not change size while the tablist does.
+      this.#resizeObserver.observe(tablist);
+    });
   }
 
   override disconnectedCallback(): void {
-    this.#sizeObserver.disconnect();
+    this.#resizeObserver.disconnect();
+    this.#mutationObserver.disconnect();
 
     super.disconnectedCallback();
-  }
-
-  override firstUpdated(): void {
-    this.#observer = new MutationObserver(this.#handleMutation);
-    this.#observer?.observe(this, OBSERVER_OPTIONS);
   }
 
   override updated(changes: PropertyValues<this>): void {
     super.updated(changes);
 
-    if (changes.has('alignment') || changes.has('vertical')) {
+    if (changes.has('alignTabs')) {
+      this.#shouldAnimate = false;
       this.#updateSelectionIndicator();
+      this.#shouldAnimate = true;
     }
-  }
 
-  #updateSlots(): void {
-    this.#setupTabs();
-    this.#setupPanels();
-  }
+    // In vertical mode, we need to observe the scroller for changes in size to
+    // determine when we need to show the menu button.
+    if (changes.has('vertical')) {
+      const scroller = this.renderRoot.querySelector('[part="scroller"]') as HTMLElement;
 
-  #onClick(): void {
-    this.listbox.togglePopover();
-  }
-
-  #onToggle = (event: ToggleEvent): void => {
-    this.#rovingTabindexController.clearElementCache();
-
-    requestAnimationFrame(() => {
-      if (!this.listbox || !this.selectedTab) {
-        return;
+      if (this.vertical) {
+        this.#resizeObserver.observe(scroller);
+      } else {
+        this.#resizeObserver.unobserve(scroller);
       }
-      this.selectedTabInListbox = this.listbox.querySelector(`#${this.selectedTab.id}`) as Tab;
-    });
-
-    if (
-      (event.newState === 'closed' && this.listbox.matches(':popover-open')) ||
-      (event.newState === 'closed' && this.listbox.matches('.\\:popover-open')) ||
-      event.newState === event.oldState
-    ) {
-      event.stopPropagation();
-      this.listbox.hidePopover();
-    }
-
-    this.#moreButton?.setAttribute('aria-expanded', isPopoverOpen(this.listbox).toString());
-  };
-
-  /**
-   * If the selected tab is selected programmatically update all the tabs.
-   */
-  #handleMutation = (mutations: MutationRecord[]): void => {
-    mutations.forEach(mutation => {
-      if (mutation.attributeName === 'selected' && mutation.oldValue === null) {
-        const selectedTab = <Tab>mutation.target;
-        this.#observer?.disconnect();
-        this.#updateSelectedTab(selectedTab);
-        this.#observer?.observe(this, OBSERVER_OPTIONS);
-      }
-    });
-  };
-
-  /**
-   * Apply accessible attributes and values to the tab buttons.
-   */
-  #setupTabs(): void {
-    const tabs = this.querySelectorAll('sl-tab');
-
-    tabs.forEach((tab, index) => {
-      tab.setAttribute('id', `${this.#tabGroupId}-tab-${index + 1}`);
-      tab.setAttribute('aria-controls', `${this.#tabGroupId}-panel-${index + 1}`);
-      tab.toggleAttribute('selected', tab === this.selectedTab);
-    });
-  }
-
-  /**
-   * Apply accessible attributes and values to the tab panels.
-   */
-  #setupPanels(): void {
-    const panels = this.querySelectorAll('sl-tab-panel');
-    const selectedPanelId = this.selectedTab?.getAttribute('aria-controls');
-    const tabIndex = this.selectedTab ? Array.from(this.querySelectorAll('sl-tab')).indexOf(this.selectedTab) : 0;
-
-    if (panels.length === 1) {
-      panels[0].setAttribute('id', `${this.#tabGroupId}-panel-${tabIndex + 1}`);
-      panels[0].setAttribute('aria-labelledby', `${this.#tabGroupId}-tab-${tabIndex + 1}`);
-      panels[0].setAttribute('aria-hidden', 'false');
-    } else {
-      panels.forEach((panel, index) => {
-        panel.setAttribute('id', `${this.#tabGroupId}-panel-${index + 1}`);
-        panel.setAttribute('aria-labelledby', `${this.#tabGroupId}-tab-${index + 1}`);
-        panel.setAttribute('aria-hidden', `${panel.getAttribute('id') !== selectedPanelId ? 'true' : 'false'}`);
-      });
     }
   }
 
-  #handleTabChange(event: Event & { target: HTMLElement }): void {
-    /**
-     * Return handler if it's not a tab
-     */
-    if (!(event.target.closest('sl-tab') instanceof Tab)) {
+  override render(): TemplateResult {
+    return html`
+      <div part="container">
+        <div part="wrapper">
+          <div class="fade-container">
+            <div class="fade fade-start"></div>
+            <div class="fade fade-end"></div>
+            <div @scroll=${this.#onScroll} part="scroller">
+              <div @click=${this.#onClick} @keydown=${this.#onKeydown} part="tablist" role="tablist">
+                <span class="indicator" role="presentation"></span>
+                <slot @slotchange=${this.#onTabSlotchange} name="tabs"></slot>
+              </div>
+            </div>
+          </div>
+
+          ${this.showMenu
+            ? html`
+                <sl-menu-button aria-label=${msg('Show all')} fill="ghost">
+                  <sl-icon name="ellipsis" slot="button"></sl-icon>
+                  ${this.menuItems?.map(
+                    menuItem => html`
+                      <sl-menu-item @click=${() => this.#onMenuItemClick(menuItem.tab)} ?disabled=${menuItem.disabled}>
+                        ${menuItem.title}
+                      </sl-menu-item>
+                    `
+                  )}
+                </sl-menu-button>
+              `
+            : nothing}
+        </div>
+      </div>
+      <div part="panels">
+        <slot @slotchange=${this.#onTabPanelSlotchange}></slot>
+      </div>
+    `;
+  }
+
+  #onClick(event: Event & { target: HTMLElement }): void {
+    const tab = event.target.closest('sl-tab');
+
+    if (!tab) {
       return;
     }
-    // Always reset the scroll when a tab is selected.
-    this.scrollTo({ top: 0 });
 
-    this.#updateSelectedTab(event.target.closest<Tab>('sl-tab') as Tab);
-    this.listbox.hidePopover();
+    this.#updateSelectedTab(tab);
+    this.#scrollToTabPanelStart();
   }
 
-  /**
-   * Update the selected tab button with attributes and values.
-   * Update the tab group state.
-   */
-  #updateSelectedTab(selectedTab: Tab): void {
-    const controls = selectedTab.getAttribute('aria-controls');
+  #onKeydown(event: KeyboardEvent & { target: HTMLElement }): void {
+    const tab = event.target.closest('sl-tab');
 
-    if (selectedTab === this.selectedTab || !controls || selectedTab.disabled) return;
+    if (tab && ['Enter', ' '].includes(event.key)) {
+      event.preventDefault();
+      event.stopPropagation();
 
-    const selectedPanel = this.querySelector(`#${controls}`);
-    const tabIndex = Array.from(this.querySelectorAll('sl-tab')).indexOf(selectedTab);
+      this.#updateSelectedTab(tab);
+      this.#scrollToTabPanelStart();
+    }
+  }
 
-    /**
-     * Reset all the selected state of the tabs, and select the clicked tab
-     */
-    this.querySelectorAll<Tab>('sl-tab').forEach((tab: Tab) => {
-      tab.removeAttribute('selected');
-      if (tab.id === selectedTab.id) {
-        tab.setAttribute('selected', '');
-        tab.focus();
-        tab.scrollIntoView({ block: 'nearest', inline: 'nearest' });
-        this.selectedTab = tab;
-      }
+  #onMenuItemClick(tab: Tab): void {
+    this.#updateSelectedTab(tab);
+  }
+
+  #onScroll(event: Event & { target: HTMLElement }): void {
+    let scrollStart = false,
+      scrollEnd = false;
+
+    if (this.vertical) {
+      const { clientHeight, scrollTop, scrollHeight } = event.target,
+        scrollable = scrollHeight > clientHeight;
+
+      scrollStart = scrollable && scrollTop > 0;
+      scrollEnd = scrollable && Math.round(scrollTop + clientHeight) < scrollHeight;
+    } else {
+      const { clientWidth, scrollLeft, scrollWidth } = event.target,
+        scrollable = scrollWidth > clientWidth;
+
+      scrollStart = scrollable && scrollLeft > 0;
+      scrollEnd = scrollable && Math.round(scrollLeft + clientWidth) < scrollWidth;
+    }
+
+    this.toggleAttribute('scroll-start', scrollStart);
+    this.toggleAttribute('scroll-end', scrollEnd);
+  }
+
+  #onTabSlotchange(event: Event & { target: HTMLSlotElement }): void {
+    this.tabs = event.target.assignedElements({ flatten: true }).filter((el): el is Tab => el instanceof Tab);
+    this.tabs.forEach((tab, index) => {
+      tab.id ||= `${this.#idPrefix}-tab-${index + 1}`;
     });
 
-    this.selectedTabInListbox = this.listbox?.querySelector(`#${selectedTab.id}`);
+    // If no tab is selected, select the first enabled one
+    this.selectedTab = this.tabs.find(tab => tab.selected) || this.tabs.find(tab => !tab.disabled);
 
-    /**
-     * Reset all the visibility of the panels,
-     * and show the panel related to the selected tab
-     */
-    const panels = this.querySelectorAll('sl-tab-panel');
-
-    if (panels.length === 1) {
-      panels[0].setAttribute('id', `${this.#tabGroupId}-panel-${tabIndex + 1}`);
-      panels[0].setAttribute('aria-labelledby', `${this.#tabGroupId}-tab-${tabIndex + 1}`);
-    } else {
-      panels.forEach(panel => {
-        panel.setAttribute('aria-hidden', `${panel !== selectedPanel ? 'true' : 'false'}`);
-      });
-    }
-
-    this.tabChange.emit(tabIndex);
-
-    this.#updateSelectionIndicator();
+    this.#rovingTabindexController.clearElementCache();
+    this.#linkTabsWithPanels();
   }
 
-  /**
-   * Handle keyboard accessible controls.
-   */
-  #handleKeydown(event: KeyboardEvent): void {
-    if (isPopoverOpen(this.listbox)) {
-      this.#rovingTabindexController.clearElementCache();
-      this.#rovingTabindexController.hostContainsFocus();
-    }
+  #onTabPanelSlotchange(event: Event & { target: HTMLSlotElement }): void {
+    this.tabPanels = event.target
+      .assignedElements({ flatten: true })
+      .filter((el): el is TabPanel => el instanceof TabPanel);
 
-    if (['Enter', ' '].includes(event.key)) {
-      event.preventDefault();
-      this.scrollTo({ top: 0 });
-      this.#updateSelectedTab(<Tab>event.target);
+    this.tabPanels.forEach((panel, index) => {
+      panel.id ||= `${this.#idPrefix}-panel-${index + 1}`;
+    });
 
-      if (isPopoverOpen(this.listbox)) {
-        this.listbox.hidePopover();
+    this.#linkTabsWithPanels();
+  }
+
+  #linkTabsWithPanels(): void {
+    this.tabs?.forEach((tab, index) => {
+      tab.toggleAttribute('selected', tab === this.selectedTab);
+
+      const panel = this.tabPanels?.at(index);
+
+      if (panel) {
+        tab.setAttribute('aria-controls', `${this.#idPrefix}-panel-${index + 1}`);
+        panel.setAttribute('aria-hidden', tab === this.selectedTab ? 'false' : 'true');
+        panel.setAttribute('aria-labelledby', `${this.#idPrefix}-tab-${index + 1}`);
+      } else {
+        tab.removeAttribute('aria-controls');
+      }
+    });
+  }
+
+  #scrollIntoViewIfNeeded(tab: Tab): void {
+    const scroller = this.renderRoot.querySelector('[part="scroller"]') as HTMLElement,
+      scrollerRect = scroller.getBoundingClientRect(),
+      tabRect = tab.getBoundingClientRect();
+
+    if (this.vertical) {
+      if (tabRect.top < scrollerRect.top) {
+        // The tab is above the top edge of the scroller
+        scroller.scrollBy({ top: tabRect.top - scrollerRect.top });
+      } else if (tabRect.bottom > scrollerRect.bottom) {
+        // The tab is below the bottom edge of the scroller
+        scroller.scrollBy({ top: tabRect.bottom - scrollerRect.bottom });
+      }
+    } else {
+      if (tabRect.left < scrollerRect.left) {
+        // The tab is to the left of the left edge of the scroller
+        scroller.scrollBy({ left: tabRect.left - scrollerRect.left });
+      } else if (tabRect.right > scrollerRect.right) {
+        // The tab is to the right of the right edge of the scroller
+        scroller.scrollBy({ left: tabRect.right - scrollerRect.right });
       }
     }
+  }
+
+  #scrollToTabPanelStart(): void {
+    const { bottom: containerBottom = 0 } =
+        this.renderRoot.querySelector('[part="container"]')?.getBoundingClientRect() || {},
+      { top: wrapperTop = 0 } = this.renderRoot.querySelector('[part="wrapper"]')?.getBoundingClientRect() || {},
+      { top = 0 } = this.renderRoot.querySelector('[part="panels"]')?.getBoundingClientRect() || {};
+
+    // Scroll to make sure the top of the panel is visible, but don't scroll too far
+    // so the tab container/wrapper may become unstuck.
+    getScrollParent(this)?.scrollBy({ top: top - (this.vertical ? wrapperTop : containerBottom) });
+  }
+
+  #updateSelectedTab(selectedTab: Tab): void {
+    if (selectedTab !== this.selectedTab) {
+      this.tabs?.forEach(tab => tab.toggleAttribute('selected', tab === selectedTab));
+
+      this.querySelectorAll('sl-tab-panel').forEach(panel => {
+        panel.setAttribute('aria-hidden', selectedTab.getAttribute('aria-controls') === panel.id ? 'false' : 'true');
+      });
+
+      this.selectedTab = selectedTab;
+      this.tabChangeEvent.emit(this.tabs?.indexOf(selectedTab) ?? 0);
+      this.#updateSelectionIndicator();
+    }
+
+    this.#scrollIntoViewIfNeeded(selectedTab);
   }
 
   #updateSelectionIndicator(): void {
-    requestAnimationFrame(() => {
-      if (!this.selectedTab || !this.selectedTabInListbox) {
-        return;
-      }
+    if (!this.selectedTab) {
+      return;
+    }
 
-      if (!this.#showMore && isPopoverOpen(this.listbox)) {
-        this.listbox.hidePopover();
-      }
+    const indicator = this.renderRoot.querySelector('.indicator') as HTMLElement,
+      tablist = this.renderRoot.querySelector('[part="tablist"]') as HTMLElement,
+      rect = this.selectedTab.getBoundingClientRect();
 
-      const axis = this.vertical ? 'Y' : 'X',
-        indicator = this.renderRoot.querySelector('.indicator') as HTMLElement,
-        wrapper = this.renderRoot.querySelector('.container') as HTMLElement,
-        tabsWrapper = this.renderRoot.querySelector('.wrapper') as HTMLElement,
-        tablist = this.renderRoot.querySelector('[role="tablist"]') as HTMLElement,
-        tabs = this.querySelectorAll('sl-tab');
+    let start = 0;
+    if (this.vertical) {
+      start = rect.top - tablist.getBoundingClientRect().top;
+    } else {
+      start = rect.left - tablist.getBoundingClientRect().left;
+    }
 
-      let totalTabsWidth = 0;
-      let totalTabsHeight = 0;
-      tabs.forEach(tab => {
-        totalTabsWidth += tab.offsetWidth;
-        totalTabsHeight += tab.offsetHeight;
+    indicator.style.transitionDuration = this.#shouldAnimate ? '' : '0s';
+
+    if (this.vertical) {
+      indicator.style.scale = `1 ${rect.height / 100}`;
+      indicator.style.translate = `0 ${start}px`;
+    } else {
+      indicator.style.scale = `${rect.width / 100} 1`;
+      indicator.style.translate = `${start}px`;
+    }
+  }
+
+  #updateSize(): void {
+    const scroller = this.renderRoot.querySelector('[part="scroller"]') as HTMLElement,
+      tablist = this.renderRoot.querySelector('[part="tablist"]') as HTMLElement;
+
+    this.showMenu = this.vertical
+      ? tablist.scrollHeight > scroller.offsetHeight
+      : tablist.scrollWidth > scroller.offsetWidth;
+
+    if (this.showMenu) {
+      this.menuItems = this.tabs?.map(tab => {
+        const title = Array.from(tab.childNodes)
+          .filter(node => node instanceof Text || (node instanceof Element && !node.slot))
+          .reduce((acc, node) => acc + node.textContent?.trim() || '', '');
+
+        const subtitle = Array.from(tab.childNodes)
+          .filter(node => node instanceof Element && node.slot === 'subtitle')
+          .reduce((acc, node) => acc + node.textContent?.trim() || '', '');
+
+        return { tab, disabled: tab.disabled, title, subtitle };
       });
+    } else {
+      this.menuItems = undefined;
+    }
 
-      this.#showMore = axis === 'X' ? totalTabsWidth > tabsWrapper.offsetWidth : totalTabsHeight > wrapper.offsetHeight;
+    this.selectedTab?.scrollIntoView();
 
-      this.requestUpdate();
-
-      let start = 0;
-      if (axis === 'X') {
-        start = this.selectedTab.getBoundingClientRect().left - tablist.getBoundingClientRect().left;
-      } else {
-        start = this.selectedTab.getBoundingClientRect().top - tablist.getBoundingClientRect().top;
-      }
-
-      // Somehow on Chromium, the offsetParent is different than on FF and Safari
-      // If on Chromium, take the `wrapper.offsetLeft` into account as well
-      if (this.selectedTab.offsetParent === wrapper) {
-        start += axis === 'X' ? wrapper.offsetLeft : wrapper.offsetTop;
-      }
-
-      indicator.style.transform = `translate${axis}(${start}px)`;
-
-      if (axis === 'X') {
-        indicator.style.removeProperty('height');
-        indicator.style.width = `${this.selectedTab.offsetWidth}px`;
-        const scrollLeft = Math.max(
-          this.selectedTab.offsetLeft + this.selectedTab.offsetWidth / 2 - tabsWrapper.clientWidth / 2,
-          0
-        );
-
-        if (scrollLeft !== tabsWrapper.scrollLeft) {
-          tabsWrapper.scrollTo({ left: scrollLeft, behavior: 'smooth' });
-        }
-      } else {
-        indicator.style.removeProperty('width');
-        indicator.style.height = `${this.selectedTab.offsetHeight}px`;
-      }
-    });
+    this.#updateSelectionIndicator();
   }
 }