+++ scripts/lib/file-headers.js
@@ -1,9 +1,0 @@
-export const legal = {
-  name: 'sl/legal',
-  fileHeader(defaultMessage) {
-    return [
-      'Copyright 2022 Sanoma Learning',
-      'SPDX-License-Identifier: Apache-2.0'
-    ];
-  }
-};