+++ packages/components/drawer/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @sl-design-system/drawer
 
+## 0.0.27
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @sl-design-system/button@0.0.27
+
+## 0.0.26
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @sl-design-system/button@0.0.26
+
+## 0.0.25
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @sl-design-system/button@0.0.25
+
+## 0.0.24
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @sl-design-system/button@0.0.24
+
 ## 0.0.23
 
 ### Patch Changes